+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2021-11-06
+
 ### Changed
 - Refactor '_vect' as 'vect' and add 'vectorize' decorator.
 - Refactor 'metric' array to 'invcov' as callable.